Refactor config to allow defining new regions (#320)
In order to let us configure and build the platform for new regions, this PR adds a new environment variable `REGION_ENV`. When set, it causes replacement configuration options to be loaded. As an example, this PR adds configurations that enable only the Meldedialog for `REGION_ENV=bonn`.

The new region can be activated with the new commands

- npm start:bonn
- npm dev:bonn
- npm build:bonn

Closes https://github.com/FixMyBerlin/fixmy.platform/issues/262

* Extend config for region-specific setup
* Add geojson boundary for sample region
* Disable subpages when their routes are not configured
* Add instance-specific configuration for reports feature
* Add build command for sample new region