chore(jest setup): only use one setup file
This simplifies our setup as it has been desired.
It makes not much sense to split up the logic running before each test into two files.
It is not 100% clear, but setupFilesAfterEnv seems to offer everything setupFiles does PLUS
globals like `expect` or `beforeAll` can be used.
See https://stackoverflow.com/questions/47587689/whats-the-difference-between-setupfiles-and-setuptestframeworkscriptfile